Move the finite abelian groups to autstr.groups, rename closure to composition
FiniteAbelianGroups sat in autstr.algebra beside the boolean algebras, but it
is a class of groups and belongs with the other two. It is also the clearest
instance of the new machinery: the advice is a separated list of cyclic orders,
which is exactly the direct-product closure of the cyclic groups, so its
docstring now points at composition.direct_product_closure. The duplicate
`_lsbf_bits` helper disappears in the move -- groups.py already had an
identical one.

`autstr.closure` becomes `autstr.composition`: the module is about composing
presentations, and "closure" already means something else in this library
(the direct-product closure of a class is one of its operations, and the
first-order closure of automatic relations is the reason any of it works).

Callers updated: the abelian benchmark, the two notebooks that import the class
(re-executed, so their outputs stay honest), and the tests, whose abelian block
moves to test_groups.py alongside the class.
