+++ autstr/groups.py
@@ -1,4 +1,11 @@
-"""Non-abelian uniformly automatic classes of finite groups.
+"""Uniformly automatic classes of finite groups.
+
+**Finite abelian groups** (`FiniteAbelianGroups`): every finite abelian group
+is a direct sum of cyclic groups, so the advice is the '#'-separated list of
+their orders in LSB-first binary and addition is blockwise. This is the
+direct-product closure of the cyclic groups, made explicit -- compare
+`autstr.composition.direct_product_closure`.
+
 
 **Groups with a cyclic subgroup of index <= 2** (`IndexTwoCyclicGroups`):
 every such group is <r, s | r^n, s^2 = r^w, s r s^-1 = r^u> with u^2 = 1
@@ -42,6 +49,7 @@ from autstr.sparse_automata import SparseDFA
 from autstr.uniform import UniformlyAutomaticClass, dfa_from_delta
 
 PAD = '*'
+SEP = '#'
 
 
 def _lsbf_bits(n: int) -> List[str]:
@@ -539,3 +547,126 @@ class ExtraspecialGroups:
 
     def get_structure(self, n: int) -> AutomaticPresentation:
         return self.cls.get_structure(self.advice(n))
+
+
+# ====================================================================
+# Finite abelian groups
+# ====================================================================
+
+class FiniteAbelianGroups:
+    """The uniformly automatic class of all finite abelian groups, presented
+    by their cyclic decompositions: the group Z_{n_1} ⊕ ... ⊕ Z_{n_k} has
+    advice bin(n_1)# ... bin(n_k)# (LSB-first binary per block)."""
+
+    def __init__(self):
+        self.sigma = {PAD, '0', '1', SEP}
+        self.cls = UniformlyAutomaticClass({
+            'U': self._universe_automaton(),
+            'A': self._addition_automaton(),
+        }, padding_symbol=PAD)
+
+    def _universe_automaton(self) -> SparseDFA:
+        """U(p, x): p is a sequence of canonical nonempty binary blocks (MSB
+        1 last, so each block denotes some n >= 1) and per block the element
+        digits denote a value < n. LSB-first comparison: the last differing
+        digit decides."""
+        states = ['between', 'pad', 'dead']
+        states += [('c', cmp, msb) for cmp in ('EQ', 'LT', 'GT') for msb in (0, 1)]
+
+        def delta(q, sym):
+            a, x = sym
+            if q == 'dead':
+                return 'dead'
+            if q == 'pad':
+                return 'pad' if (a, x) == (PAD, PAD) else 'dead'
+            if a in ('0', '1') and x in ('0', '1'):
+                cmp = q[1] if q != 'between' else 'EQ'
+                if x != a:
+                    cmp = 'LT' if x == '0' else 'GT'
+                return ('c', cmp, int(a))
+            if (a, x) == (SEP, SEP):
+                # block ends: canonical (msb 1) and element value < n
+                if q != 'between' and q[1] == 'LT' and q[2] == 1:
+                    return 'between'
+                return 'dead'
+            if (a, x) == (PAD, PAD) and q == 'between':
+                return 'pad'
+            return 'dead'
+
+        return dfa_from_delta(self.sigma, states, 2, delta, 'between', {'between', 'pad'})
+
+    def _addition_automaton(self) -> SparseDFA:
+        """A(p, x, y, z): per block, x + y = z + b*n for some b in {0,1}
+        (i.e. z = x + y mod n), checked digit by digit LSB-first with one
+        signed carry per branch b."""
+        carries = list(range(-2, 3)) + ['F']
+        states = [('a', c0, c1) for c0 in carries for c1 in carries]
+        states += ['pad', 'dead']
+
+        def step(carry, n_bit, x, y, z, b):
+            if carry == 'F':
+                return 'F'
+            s = carry + x + y - z - b * n_bit
+            if s % 2 != 0:
+                return 'F'
+            s //= 2
+            return s if -2 <= s <= 2 else 'F'
+
+        def delta(q, sym):
+            a, x, y, z = sym
+            if q == 'dead':
+                return 'dead'
+            if q == 'pad':
+                return 'pad' if all(s == PAD for s in sym) else 'dead'
+            c0, c1 = q[1], q[2]
+            if a in ('0', '1') and all(s in ('0', '1') for s in (x, y, z)):
+                bits = (int(a), int(x), int(y), int(z))
+                return ('a', step(c0, *bits, 0), step(c1, *bits, 1))
+            if all(s == SEP for s in sym):
+                # block ends: some branch must close with carry 0
+                return ('a', 0, 0) if 0 in (c0, c1) else 'dead'
+            if all(s == PAD for s in sym) and (c0, c1) == (0, 0):
+                return 'pad'
+            return 'dead'
+
+        return dfa_from_delta(self.sigma, states, 4, delta, ('a', 0, 0),
+                              {('a', 0, 0), 'pad'})
+
+    # ---------------- encodings and class operations ----------------
+
+    def advice(self, orders: Sequence[int]) -> List[str]:
+        """Advice string of Z_{n_1} ⊕ ... ⊕ Z_{n_k}."""
+        word = []
+        for n in orders:
+            if n < 1:
+                raise ValueError("cyclic factor orders must be >= 1")
+            word += _lsbf_bits(n) + [SEP]
+        return word
+
+    def encode(self, element: Union[int, Sequence[int]], orders: Sequence[int]) -> List[str]:
+        """Encode a group element (one value per cyclic factor)."""
+        if isinstance(element, int):
+            element = (element,)
+        if len(element) != len(orders):
+            raise ValueError(f"element has {len(element)} components, group has {len(orders)}")
+        word = []
+        for a, n in zip(element, orders):
+            if not 0 <= a < n:
+                raise ValueError(f"component {a} not in Z_{n}")
+            block_len = len(_lsbf_bits(n))
+            digits = _lsbf_bits(a) if a > 0 else ['0']
+            word += digits + ['0'] * (block_len - len(digits)) + [SEP]
+        return word
+
+    def evaluate(self, phi) -> Tuple[SparseDFA, List[str]]:
+        return self.cls.evaluate(phi)
+
+    def check(self, phi, orders: Sequence[int], **elements) -> bool:
+        """Model check against Z_{n_1} ⊕ ... ⊕ Z_{n_k}; free variables can be
+        assigned group elements (tuples with one value per factor, or a
+        single int for a cyclic group)."""
+        words = {name: self.encode(e, orders) for name, e in elements.items()}
+        return self.cls.check(phi, self.advice(orders), **words)
+
+    def get_structure(self, orders: Sequence[int]) -> AutomaticPresentation:
+        return self.cls.get_structure(self.advice(orders))
